refactor sublists to support bs

+++ src/routes/listbuilder/components/SublistComponents/EditSublistModal.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
 	import type { List, Sublist } from "$lib/types/list.svelte";
 	import { Dialog } from "$lib/generic";
+	import { getBSCbyId } from "$lib/data/battlefieldSupport";
 
 	type Props = {
 		list: List;
@@ -26,7 +27,7 @@
 
 	function cancelSublist() {
 		let existingSublist = list.getSublist(sublist!.id);
-		if (existingSublist && existingSublist.checked.length == 0) {
+		if (existingSublist && existingSublist.checked.length == 0 && existingSublist.checkedBS.size == 0) {
 			list.deleteSublist(existingSublist.id);
 		}
 		open = false;
@@ -35,10 +36,11 @@
 	function updateSublist() {
 		let existingSublist = list.getSublist(sublist!.id);
 		if (existingSublist && sublist) {
-			if (sublist?.checked.length == 0) {
+			if (sublist?.checked.length == 0 && sublist?.checkedBS.size == 0) {
 				list.deleteSublist(sublist.id);
 			} else {
 				existingSublist.checked = sublist.checked;
+				existingSublist.checkedBS = sublist.checkedBS;
 			}
 		}
 		open = false;
@@ -50,7 +52,7 @@
 		<div class="edit-sublist-modal-content">
 			<div class="edit-sublist-units">
 				<div></div>
-				<p class="center">Unit</p>
+				<p>Unit</p>
 				<p class="center">Skill</p>
 				<p class="center">PV</p>
 				{#each list.units.toSorted((a, b) => {
@@ -75,6 +77,30 @@
 					<p class="center">{unit.cost}</p>
 				{/each}
 			</div>
+			<div class="edit-sublists-bs">
+				<div></div>
+				<p>Battlefield Support</p>
+				<p class="center">BSP</p>
+				<p class="center">PV</p>
+				{#each list.bsList.entries() as [id, count], index}
+					{@const bsData = getBSCbyId(id)}
+					<div class="inline">
+						<input type="number" min="0" max={count} bind:value={() => sublist?.checkedBS.get(id) ?? 0, (v) => sublist?.checkedBS.set(id, v)} />
+						<p>/{count}x</p>
+					</div>
+					<label for={`checkbox${index}`}>{bsData?.name}</label>
+					{#if bsData?.bspCost}
+						<p class="center">{bsData.bspCost}({bsData.bspCost * count})</p>
+					{:else}
+						<p class="center">-</p>
+					{/if}
+					{#if bsData?.pvCost}
+						<p class="center">{bsData.pvCost}({bsData.pvCost * count})</p>
+					{:else}
+						<p class="center">-</p>
+					{/if}
+				{/each}
+			</div>
 			<div class="edit-sublist-stats">
 				<p><span class="muted">Units:</span> {`${count}${list.options?.sublistMaxUnits ? `/${list.options.sublistMaxUnits}` : ""}`}</p>
 				<p><span class="muted">PV:</span> {`${pv}${list.options?.sublistMaxPv ? `/${list.options.sublistMaxPv}` : ""}`}</p>
@@ -96,7 +122,8 @@
 		grid-template-rows: 1fr max-content max-content;
 		gap: 16px;
 	}
-	.edit-sublist-units {
+	.edit-sublist-units,
+	.edit-sublists-bs {
 		display: grid;
 		grid-template-columns: max-content 1fr max-content max-content;
 		grid-auto-rows: max-content;
